requirements.txt
bokeh~=3.0.3
discord~=2.1.0
nltk~=3.8
python-dotenv~=0.21.0
requests~=2.28.1
emoji~=2.2.0
aiohttp~=3.8.3
pandas~=1.5.2